migrate from using backend instances to using GAE modules.
