+++ cloudpush-ear/src/main/application/META-INF/appengine-application.xml
@@ -1,0 +1,4 @@
+<?xml version="1.0" encoding="utf-8" standalone="no"?>
+<appengine-application xmlns="http://appengine.google.com/ns/1.0">
+<application>ENTER_YOUR_APP_ID_HERE</application>
+</appengine-application>
