+++ cloudpush-front/src/main/java/com/google/cloud/solutions/mobilepushnotification/DeviceCleanupServlet.java
