+++ cloudpush-front/src/main/java/com/google/cloud/solutions/mobilepushnotification/dummybackendsample/DummyGameEndpoint.java
