+++ cloudpush-front/src/main/webapp/WEB-INF/appengine-web.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="utf-8"?>
+<appengine-web-app xmlns="http://appengine.google.com/ns/1.0">
+<application>ENTER_YOUR_APP_ID_HERE</application>
+  <module>default</module>
+  <version>1</version>
+
+  <!--
+    Allows App Engine to send multiple requests to one instance in parallel:
+  -->
+  <threadsafe>true</threadsafe>
+
+  <!-- Configure java.util.logging -->
+  <system-properties>
+    <property name="java.util.logging.config.file" value="WEB-INF/logging.properties"/>
+  </system-properties>
+
+  <!--
+    HTTP Sessions are disabled by default. To enable HTTP sessions specify:
+
+      <sessions-enabled>true</sessions-enabled>
+
+    It's possible to reduce request latency by configuring your application to
+    asynchronously write HTTP session data to the datastore:
+
+      <async-session-persistence enabled="true" />
+
+    With this feature enabled, there is a very small chance your app will see
+    stale session data. For details, see
+    http://code.google.com/appengine/docs/java/config/appconfig.html#Enabling_Sessions
+  -->
+
+</appengine-web-app>
