+++ cloudpush-front/src/main/webapp/WEB-INF/queue.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<queue-entries>
+  <queue>
+    <name>notification-preprocessing</name>
+    <rate>200/s</rate>
+    <bucket-size>100</bucket-size>
+    <target>default</target>
+  </queue>
+  <queue>
+    <name>notification-device-token-cleanup</name>
+    <rate>10/s</rate>
+    <bucket-size>20</bucket-size>
+    <target>default</target>
+  </queue>
+  <queue>
+    <name>notification-delivery</name>
+    <mode>pull</mode>
+  </queue>
+</queue-entries>
