+++ cloudpush-worker/lib/README.txt
@@ -1,0 +1,6 @@
+Please put the following two libraries here:
+- JavaPNS_2.2.jar
+- bcprov-jdk15-146.jar
+
+If you download newer version of the library, you should update the pom.xml
+file to reflect the change.
