+++ cloudpush-worker/pom.xml
@@ -1,0 +1,123 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.google.cloud.solutions.mobilepushnotification</groupId>
+    <artifactId>solutions-ios-push-notification-sample-java</artifactId>
+    <version>1.0</version>
+  </parent>
+
+  <groupId>com.google.cloud.solutions.mobilepushnotification</groupId>
+  <artifactId>cloudpush-worker</artifactId>
+  <packaging>war</packaging>
+  <version>1.0</version>
+
+  <name>cloudpush-worker</name>
+
+  <dependencies>
+    <!-- Compile/runtime dependencies -->
+    <dependency>
+        <groupId>com.google.appengine</groupId>
+        <artifactId>appengine-api-1.0-sdk</artifactId>
+        <version>${appengine.target.version}</version>
+    </dependency>
+    <dependency>
+        <groupId>javax.servlet</groupId>
+        <artifactId>servlet-api</artifactId>
+        <version>2.5</version>
+        <scope>provided</scope>
+    </dependency>
+    <dependency>
+        <groupId>javax.jdo</groupId>
+        <artifactId>jdo-api</artifactId>
+        <version>3.0.1</version>
+    </dependency>
+    <dependency>
+        <groupId>net.sf.jsr107cache</groupId>
+        <artifactId>jsr107cache</artifactId>
+        <version>1.1</version>
+    </dependency>
+
+    <!-- JavaPNS dependency -->
+    <dependency>
+      <groupId>ios-push-library</groupId>
+      <artifactId>JavaPNS</artifactId>
+      <version>2.2</version>
+      <scope>system</scope>
+      <systemPath>${basedir}/lib/JavaPNS_2.2.jar</systemPath>
+    </dependency>
+    <dependency>
+      <groupId>ios-push-library</groupId>
+      <artifactId>bcprov-jdk15</artifactId>
+      <version>1.46</version>
+      <scope>system</scope>
+      <systemPath>${basedir}/lib/bcprov-jdk15-146.jar</systemPath>
+    </dependency>
+    <dependency>
+      <groupId>log4j</groupId>
+      <artifactId>log4j</artifactId>
+      <version>1.2.14</version>
+    </dependency>
+
+    <!--  Gson: Java to Json conversion -->
+    <dependency>
+      <groupId>com.google.code.gson</groupId>
+      <artifactId>gson</artifactId>
+      <version>2.3.1</version>
+      <scope>compile</scope>
+    </dependency>
+  </dependencies>
+
+  <build>
+    <pluginManagement>
+      <plugins>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <version>2.5.1</version>
+          <artifactId>maven-compiler-plugin</artifactId>
+          <configuration>
+            <source>1.7</source>
+            <target>1.7</target>
+          </configuration>
+        </plugin>
+        <plugin>
+          <groupId>org.apache.maven.plugins</groupId>
+          <artifactId>maven-war-plugin</artifactId>
+          <version>2.3</version>
+          <configuration>
+            <archiveClasses>false</archiveClasses>
+            <webResources>
+              <!-- in order to interpolate version from pom into appengine-web.xml -->
+              <resource>
+                <directory>${basedir}/src/main/webapp/WEB-INF</directory>
+                <filtering>true</filtering>
+                <targetPath>WEB-INF</targetPath>
+              </resource>
+              <resource>
+                <directory>${basedir}/lib</directory>
+                <filtering>false</filtering>
+                <targetPath>WEB-INF/lib</targetPath>
+              </resource>
+            </webResources>
+          </configuration>
+        </plugin>
+        <plugin>
+          <groupId>com.google.appengine</groupId>
+          <artifactId>appengine-maven-plugin</artifactId>
+          <version>${appengine.target.version}</version>
+          <configuration>
+            <enableJarClasses>true</enableJarClasses>
+          </configuration>
+          <executions>
+            <execution>
+              <phase>compile</phase>
+              <goals>
+                <goal>enhance</goal>
+              </goals>
+            </execution>
+          </executions>
+        </plugin>
+      </plugins>
+    </pluginManagement>
+  </build>
+</project>
