+++ cloudpush-worker/src/main/webapp/WEB-INF/web.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="utf-8" standalone="no"?>
+<web-app  xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
+  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="2.5"
+  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
+
+  <servlet>
+    <servlet-name>PushNotificationWorkerServlet</servlet-name>
+    <servlet-class>com.google.cloud.solutions.mobilepushnotification.PushNotificationWorkerServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>PushNotificationWorkerServlet</servlet-name>
+    <url-pattern>/_ah/start</url-pattern>
+  </servlet-mapping>
+
+
+  <servlet>
+    <servlet-name>PushFeedbackProcessingServlet</servlet-name>
+    <servlet-class>com.google.cloud.solutions.mobilepushnotification.PushFeedbackProcessingServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>PushFeedbackProcessingServlet</servlet-name>
+    <url-pattern>/admin/push/feedbackservice/process</url-pattern>
+  </servlet-mapping>
+
+  <servlet>
+    <servlet-name>NotificationCleanupServlet</servlet-name>
+    <servlet-class>com.google.cloud.solutions.mobilepushnotification.NotificationCleanupServlet</servlet-class>
+  </servlet>
+  <servlet-mapping>
+    <servlet-name>NotificationCleanupServlet</servlet-name>
+    <url-pattern>/admin/push/notifications/cleanup</url-pattern>
+  </servlet-mapping>
+
+  <security-constraint>
+    <web-resource-collection>
+      <url-pattern>/admin/*</url-pattern>
+    </web-resource-collection>
+    <auth-constraint>
+      <role-name>admin</role-name>
+    </auth-constraint>
+  </security-constraint>
+</web-app>
