+++ pom.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <groupId>com.google.cloud.solutions.mobilepushnotification</groupId>
+  <artifactId>solutions-ios-push-notification-sample-java</artifactId>
+  <version>1.0</version>
+  <packaging>pom</packaging>
+  <name>solutions-ios-push-notification-sample-java</name>
+
+  <properties>
+    <appengine.app.version>1</appengine.app.version>
+    <appengine.target.version>1.9.21</appengine.target.version>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+  </properties>
+  <prerequisites>
+    <maven>3.1.0</maven>
+  </prerequisites>
+  <modules>
+    <module>cloudpush-ear</module>
+    <module>cloudpush-front</module>
+    <module>cloudpush-worker</module>
+  </modules>
+</project>
