--- src/com/google/cloud/solutions/mobilepushnotification/EMF.java
@@ -1,31 +1,0 @@
-/*
- * Copyright (c) 2013 Google Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
- * in compliance with the License. You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- */
-package com.google.cloud.solutions.mobilepushnotification;
-
-import javax.persistence.EntityManagerFactory;
-import javax.persistence.Persistence;
-
-/**
- * Holds a singleton EntityManagerFactory
- */
-public final class EMF {
-  private static final EntityManagerFactory emfInstance =
-      Persistence.createEntityManagerFactory("transactions-optional");
-
-  private EMF() {}
-
-  public static EntityManagerFactory get() {
-    return emfInstance;
-  }
-}
