--- war/WEB-INF/backends.xml
@@ -1,4 +1,0 @@
-<backends>
-  <backend name="worker">
-  </backend>
-</backends>
