--- war/WEB-INF/web.xml
@@ -1,81 +1,0 @@
-<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<web-app xmlns="http://java.sun.com/xml/ns/javaee" xmlns:web="http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="2.5"
-  xsi:schemaLocation="http://java.sun.com/xml/ns/javaee http://java.sun.com/xml/ns/javaee/web-app_2_5.xsd">
-
-  <servlet>
-    <servlet-name>SystemServiceServlet</servlet-name>
-    <servlet-class>com.google.api.server.spi.SystemServiceServlet</servlet-class>
-    <init-param>
-      <param-name>services</param-name>
-      <param-value>com.google.cloud.solutions.mobilepushnotification.DeviceRegistrationEndpoint,com.google.cloud.solutions.mobilepushnotification.dummybackendsample.DummyGameEndpoint</param-value>
-    </init-param>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>SystemServiceServlet</servlet-name>
-    <url-pattern>/_ah/spi/*</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>PushNotificationWorkerServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.PushNotificationWorkerServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>PushNotificationWorkerServlet</servlet-name>
-    <url-pattern>/_ah/start</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>PushPreProcessingServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.PushPreProcessingServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>PushPreProcessingServlet</servlet-name>
-    <url-pattern>/admin/push/preprocessing</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>PushFeedbackProcessingServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.PushFeedbackProcessingServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>PushFeedbackProcessingServlet</servlet-name>
-    <url-pattern>/admin/push/feedbackservice/process</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>DeviceCleanupServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.DeviceCleanupServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>DeviceCleanupServlet</servlet-name>
-    <url-pattern>/admin/push/device/cleanup</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>NotificationCleanupServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.NotificationCleanupServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>NotificationCleanupServlet</servlet-name>
-    <url-pattern>/admin/push/notifications/cleanup</url-pattern>
-  </servlet-mapping>
-
-  <servlet>
-    <servlet-name>DummyGameAdminServlet</servlet-name>
-    <servlet-class>com.google.cloud.solutions.mobilepushnotification.dummybackendsample.DummyGameAdminServlet</servlet-class>
-  </servlet>
-  <servlet-mapping>
-    <servlet-name>DummyGameAdminServlet</servlet-name>
-    <url-pattern>/admin/test</url-pattern>
-  </servlet-mapping>
-
-  <security-constraint>
-    <web-resource-collection>
-      <url-pattern>/admin/*</url-pattern>
-    </web-resource-collection>
-    <auth-constraint>
-      <role-name>admin</role-name>
-    </auth-constraint>
-  </security-constraint>
-</web-app>
